// src/proto/types/campaigns/Campaign.ts
export interface Campaign__Output {
  'id'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
  'createdBy'?: (string);
  'updatedBy'?: (string);
  'state'?: (string);
  'title'?: (string);
  'description'?: (string);
  'launchDate'?: (string);
  'endDate'?: (string);
  'minEquity'?: (number);
  'minAmount'?: (number);
  'valuationCap'?: (number);
  'discount'?: (number);
  'fundingGoal'?: (number);
  'round'?: (string);
  'pitchTitle'?: (string);
  'pitchContent'?: (string);
  'companyId'?: (number);
  'recentNews': (_recentNews_RecentNew__Output)[];
  'company'?: (_companies_Company__Output | null);
  'type'?: (string);
  'category'?: (string);
  'documents': (_documents_Document__Output)[];
  'faq': (_faq_FAQ__Output)[];
  'code'?: (string);
  'companyPicture'?: (string);
  'totalInvestor'?: (number);
  'bannerPicture'?: (string);
  'countryId'?: (number);
  'countryName'?: (string);
  'cityId'?: (number);
  'cityName'?: (string);
  'bookmarks': (_userMapping_UserMapping__Output)[];
  'hashtags': (_campaigns_HashTags__Output)[];
  '_id': "id";
  '_createdAt': "createdAt";
  '_updatedAt': "updatedAt";
  '_createdBy': "createdBy";
  '_updatedBy': "updatedBy";
  '_state': "state";
  '_title': "title";
  '_description': "description";
  '_launchDate': "launchDate";
  '_endDate': "endDate";
  '_minEquity': "minEquity";
  '_minAmount': "minAmount";
  '_valuationCap': "valuationCap";
  '_discount': "discount";
  '_fundingGoal': "fundingGoal";
  '_round': "round";
  '_pitchTitle': "pitchTitle";
  '_pitchContent': "pitchContent";
  '_companyId': "companyId";
  '_company': "company";
  '_type': "type";
  '_category': "category";
  '_code': "code";
  '_companyPicture': "companyPicture";
  '_totalInvestor': "totalInvestor";
  '_bannerPicture': "bannerPicture";
  '_countryId': "countryId";
  '_countryName': "countryName";
  '_cityId': "cityId";
  '_cityName': "cityName";
}