// src/proto/types/campaigns/QueryCampaign.ts
// Original file: src/proto/proto/campaign.proto


export interface QueryCampaign {
  'id'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
  'createdBy'?: (string);
  'updatedBy'?: (string);
  'state'?: (string);
  'page'?: (number);
  'limit'?: (number);
  'sortBy'?: (string);
  'sortOrder'?: (string);
  'title'?: (string);
  'description'?: (string);
  'launchDate'?: (string);
  'endDate'?: (string);
  'minEquity'?: (number);
  'minAmount'?: (number | string);
  'valuationCap'?: (number | string);
  'discount'?: (number);
  'fundingGoal'?: (number | string);
  'round'?: (string);
  'pitchTitle'?: (string);
  'pitchContent'?: (string);
  'companyId'?: (number);
  'cityId'?: (number);
  'type'?: (string);
  'category'?: (string);
  '_id'?: "id";
  '_createdAt'?: "createdAt";
  '_updatedAt'?: "updatedAt";
  '_createdBy'?: "createdBy";
  '_updatedBy'?: "updatedBy";
  '_state'?: "state";
  '_page'?: "page";
  '_limit'?: "limit";
  '_sortBy'?: "sortBy";
  '_sortOrder'?: "sortOrder";
  '_title'?: "title";
  '_description'?: "description";
  '_launchDate'?: "launchDate";
  '_endDate'?: "endDate";
  '_minEquity'?: "minEquity";
  '_minAmount'?: "minAmount";
  '_valuationCap'?: "valuationCap";
  '_discount'?: "discount";
  '_fundingGoal'?: "fundingGoal";
  '_round'?: "round";
  '_pitchTitle'?: "pitchTitle";
  '_pitchContent'?: "pitchContent";
  '_companyId'?: "companyId";
  '_cityId'?: "cityId";
  '_type'?: "type";
  '_category'?: "category";
}

export interface QueryCampaign__Output {
  'id'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
  'createdBy'?: (string);
  'updatedBy'?: (string);
  'state'?: (string);
  'page'?: (number);
  'limit'?: (number);
  'sortBy'?: (string);
  'sortOrder'?: (string);
  'title'?: (string);
  'description'?: (string);
  'launchDate'?: (string);
  'endDate'?: (string);
  'minEquity'?: (number);
  'minAmount'?: (number);
  'valuationCap'?: (number);
  'discount'?: (number);
  'fundingGoal'?: (number);
  'round'?: (string);
  'pitchTitle'?: (string);
  'pitchContent'?: (string);
  'companyId'?: (number);
  'cityId'?: (number);
  'type'?: (string);
  'category'?: (string);
  '_id': "id";
  '_createdAt': "createdAt";
  '_updatedAt': "updatedAt";
  '_createdBy': "createdBy";
  '_updatedBy': "updatedBy";
  '_state': "state";
  '_page': "page";
  '_limit': "limit";
  '_sortBy': "sortBy";
  '_sortOrder': "sortOrder";
  '_title': "title";
  '_description': "description";
  '_launchDate': "launchDate";
  '_endDate': "endDate";
  '_minEquity': "minEquity";
  '_minAmount': "minAmount";
  '_valuationCap': "valuationCap";
  '_discount': "discount";
  '_fundingGoal': "fundingGoal";
  '_round': "round";
  '_pitchTitle': "pitchTitle";
  '_pitchContent': "pitchContent";
  '_companyId': "companyId";
  '_cityId': "cityId";
  '_type': "type";
  '_category': "category";
}